ptb/requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.6
requests==2.31.0
lxml!=4.9.3
ply==3.11
boto3==1.34.11